main_app/pyproject.toml
[tool.poetry]
name = "main-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
authlib = "^1.3.2"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
uvicorn = "^0.32.1"
ruff = "^0.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
